Fat pointer to RealSexp implements SliceView
